+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### next
 
-* TODO: Replace this bullet point with an actual description of a change.
+* Added a monkey-patch for the constellation Karafka 1.4 and Thor 1.3,
+  on Ruby >=2.7 (#42)
 
 ### 1.7.5